src/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
Django==3.1.7
django-autoslug==1.9.8
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient @ file:///C:/Users/Nitro5/Desktop/PycharmProjects/tryDjangoREST/src/mysqlclient-1.4.6-cp37-cp37m-win32.whl
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
